% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disturbance-analysis.R
\name{DisturbancePercentCover}
\alias{DisturbancePercentCover}
\title{Disturbance cover by disturbance type and transect}
\usage{
DisturbancePercentCover(conn, path.to.data, park, spring, field.season,
  drop.zeroes = FALSE, data.source = "database")
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{spring}{Optional. Spring code to filter on, e.g. "LAKE_P_BLUE0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{drop.zeroes}{Should percent covers of 0 be included in the dataset? Defaults to FALSE.}

\item{data.source}{Character string indicating whether to access data in the spring veg database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}
}
\value{
A tibble with columns Park, SpringCode, SpringName, FieldSeason, TransectNumber, CanopyPercentCover
}
\description{
Disturbance cover by disturbance type and transect
}
\details{
Only includes data from visits labeled 'Primary.' Percent cover is calculated only from points at which data were collected - if only three points were recorded for a transect and one of them had bike disturbance, the disturbance type BK would have 33.3% cover for that transect.
}
