% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{FormatPlot}
\alias{FormatPlot}
\title{Apply some standard formatting to a ggplot object.}
\usage{
FormatPlot(p, spring, spring.name, field.seasons, sample.sizes, plot.title,
  sub.title = "default", rotate.x.labs = FALSE, ymax = FALSE,
  ymin = FALSE, xmax = FALSE, xmin = FALSE)
}
\arguments{
\item{p}{A ggplot object.}

\item{spring}{The spring code.}

\item{spring.name}{The spring name.}

\item{field.seasons}{Either a single field season name, or a vector of field season names.}

\item{sample.sizes}{A dataframe with columns SpringCode, FieldSeason, NTransects (i.e. sample size).}

\item{plot.title}{The title of the plot.}

\item{sub.title}{Optional custom plot subtitle.}

\item{rotate.x.labs}{Boolean indicating whether to rotate x axis labels 90 degrees.}

\item{ymax}{Optional maximum y limit.}

\item{ymin}{Optional minimum y limit.}

\item{xmax}{Optional maximum x limit.}

\item{xmin}{Optional minimum x limit.}
}
\value{
A ggplot object.
}
\description{
Apply some standard formatting to a ggplot object.
}
