% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetDisturbanceTypes}
\alias{GetDisturbanceTypes}
\title{List disturbance types monitored as part of the spring veg protocol.}
\usage{
GetDisturbanceTypes()
}
\value{
A tibble with columns Code, Label, Summary, and SortOrder
}
\description{
List disturbance types monitored as part of the spring veg protocol.
}
\details{
This data is saved as internal data in R/sysdata.rda. If the codes are modified in the database, the data in the package will need to be manually updated.
}
