% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetSampleSizes}
\alias{GetSampleSizes}
\title{Compute sample size by spring and field season}
\usage{
GetSampleSizes(data)
}
\arguments{
\item{data}{A data frame with at least the following columns: SpringCode, FieldSeason, TransectNumber.}
}
\value{
A dataframe with columns for spring code, field season, and number of transects (i.e. sample size).
}
\description{
Compute sample size by spring and field season
}
