% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetSpringName}
\alias{GetSpringName}
\title{Get the name of a spring from the spring code}
\usage{
GetSpringName(conn, path.to.data, spring.code, data.source = "database")
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{data.source}{Character string indicating whether to access data in the spring veg database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}

\item{spring}{Spring code to get the name for, e.g. "LAKE_P_BLUE0".}
}
\value{
The name of the spring
}
\description{
Get the name of a spring from the spring code
}
