% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-figures.R
\name{HistogramCanopyPercentCover}
\alias{HistogramCanopyPercentCover}
\title{Histogram of LPI canopy cover}
\usage{
HistogramCanopyPercentCover(conn, path.to.data, spring, field.season,
  data.source = "database", plot.title, sub.title, x.lab, y.lab,
  breaks = seq(0, 100, by = 10), ymax)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{spring}{Spring code to generate a plot for, e.g. "LAKE_P_BLUE0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the spring veg database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}

\item{plot.title}{Optional custom plot title. Leave blank to use a sensible default. Use "" to omit the title.}

\item{sub.title}{Optional custom plot subtitle.  Leave blank to use a sensible default. Use "" to omit the subtitle.}

\item{x.lab}{Optional X axis label. Leave blank to use a sensible default. Use "" to omit the axis label.}

\item{y.lab}{Y axis label. Leave blank to use a sensible default. Use "" to omit the axis label.}

\item{breaks}{Optional numeric vector giving histogram bin boundaries. Defaults to \code{seq(0, 100, by = 10)} to give 10 bins of width 10.}

\item{ymax}{Optional maximum y limit.}
}
\value{
A ggplot object.
}
\description{
Histogram of LPI canopy cover
}
\details{
Only includes data from visits labeled 'Primary.' Counts both plant canopy and other canopy (e.g. litter) as cover. Does not take into account the number of canopy layers, only presence/absence of canopy at a point. I.e., a point with a single plant counts the same as a point with multiple plants in multiple strata. Percent cover is calculated only from points at which data were collected - if only three points were recorded for a transect and one of them had canopy cover, that transect would have 33.3% cover.
}
