% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi-qc.R
\name{LpiQcMissingPoints}
\alias{LpiQcMissingPoints}
\title{Check for missing LPI points}
\usage{
LpiQcMissingPoints(conn, path.to.data, park, spring, field.season,
  data.source = "database", points = seq(0.5, 15, 0.5))
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{spring}{Optional. Spring code to filter on, e.g. "LAKE_P_BLUE0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the spring veg database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}

\item{points}{Vector representing the expected points on a transect. Defaults to \code{seq(0.5, 15, 0.5)} (a 15m transect sampled every 0.5m and starting at 0.5m).}
}
\value{
None.
}
\description{
Check for missing LPI points
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    LpiQcMissingPoints(conn, park = "LAKE")
    LpiQcMissingPoints(conn, spring = "LAKE_P_BLUE0", field.season = "2019")
    LpiQcMissingPoints(path.to.data = "path/to/data", data.source = "local")
    CloseDatabaseConnection(conn)
}
}
