% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lpi-qc.R
\name{LpiQcUNKSpecies}
\alias{LpiQcUNKSpecies}
\title{List occurrences of "UNK" unknown plant species in LPI data}
\usage{
LpiQcUNKSpecies(conn, path.to.data, park, spring, field.season,
  data.source = "database")
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{spring}{Optional. Spring code to filter on, e.g. "LAKE_P_BLUE0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the spring veg database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}
}
\value{
A tibble with columns Park, SpringCode, SpringName, VisitType, FieldSeason, StartDate, UnknownPlantCode, TransectNumber, LocationOnTape_m. Note that LocationOnTape_m is a comma-separated list of each point where the unknown species occurs.
}
\description{
List occurrences of "UNK" unknown plant species in LPI data
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    LpiQcUNKSpecies(conn, park = "LAKE")
    LpiQcUNKSpecies(conn, spring = "LAKE_P_BLUE0", field.season = "2019")
    LpiQcUNKSpecies(path.to.data = "path/to/data", data.source = "local")
    CloseDatabaseConnection(conn)
}
}
