% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{OpenDatabaseConnection}
\alias{OpenDatabaseConnection}
\title{Open a Connection to the Spring Veg Database}
\usage{
OpenDatabaseConnection(use.mojn.default = TRUE, drv = odbc::odbc(),
  ...)
}
\arguments{
\item{use.mojn.default}{Connect to the MOJN Spring Veg database? MOJN staff should use this option. Defaults to \code{TRUE}.}

\item{drv}{DBI driver to use. Defaults to \code{odbc::odbc()}.}

\item{...}{Additional arguments to \code{\link[pool]{dbPool}}. Ignored if \code{use.mojn.default} is \code{TRUE}.}
}
\value{
A database connection pool object
}
\description{
Open a Connection to the Spring Veg Database
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
}
}
