% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-figures.R
\name{TableSpeciesPerTransect}
\alias{TableSpeciesPerTransect}
\title{Table of species detected through LPI vs veg inventory}
\usage{
TableSpeciesPerTransect(conn, path.to.data, park, spring, field.season,
  dt.options, data.source = "database", ...)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{spring}{Optional. Spring code to generate a plot for, e.g. "LAKE_P_BLUE0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{dt.options}{Optional. Options to be passed to DT::datatable().}

\item{data.source}{Character string indicating whether to access data in the spring veg database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}

\item{...}{Optional. Additional arguments to be passed to DT::datatable().}
}
\value{
An HTML widget.
}
\description{
Table of species detected through LPI vs veg inventory
}
\details{
Omits TBD and UNK species from counts. Only includes data from visits labeled 'Primary.'
}
