% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegetation-analysis.R
\name{WaterPercentCover}
\alias{WaterPercentCover}
\title{Water percent cover by transect}
\usage{
WaterPercentCover(conn, path.to.data, park, spring, field.season,
  data.source = "database")
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{spring}{Optional. Spring code to filter on, e.g. "LAKE_P_BLUE0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the spring veg database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}
}
\value{
A tibble with columns Park, SpringCode, SpringName, FieldSeason, TransectNumber, WaterPercentCover
}
\description{
Water percent cover by transect
}
\details{
Only includes data from visits labeled 'Primary.' Points with water recorded as "NA" and "ND" are omitted. Omits Blue Point transects 1 - 3 in 2019 since water presence/absence was not yet being recorded consistently.
}
